Suppress some warnings (Windows)

+++ core/src/MultiFormatReader.cpp
@@ -78,7 +78,7 @@ Results MultiFormatReader::readMultiple(const BinaryBitmap& image, int maxSymbol
 			auto it = std::remove_if(res.begin(), res.end(), [](auto&& r) { return !r.isValid(); });
 			res.erase(it, res.end());
 		}
-		maxSymbols -= r.size();
+		maxSymbols -= (int) r.size();
 		res.insert(res.end(), std::move_iterator(r.begin()), std::move_iterator(r.end()));
 		if (maxSymbols <= 0)
 			break;
